Added the ability to zoom into a tune at a point
Type 1 or 2 to run the tune1 or tune2 algorithm. The scale on the sides will change to numbers, type the y and x coordinates and enter.
At that point the view changes to SWRs in the L and C space. It is zoomable and scrollable (with the letters as arrow keys as shown). Press C to switch LC to CL. Press H to switch to hit mode and see where the tune algorithm 'looked' into the L and C space.