Fixed capacitor map, added relocate feature
Relocate feature now allows browsing the resistance/reactance space while viewing the LC map. This really gives a good idea of what's going on there. Type r to toggle the mode. When in relocate, the scroll keys change to relocate keys. Observe the complex impedance at the top for location.